complete notes page screen
