+++ src/components/EditorContainer.js/editor-container.module.css
@@ -1,20 +1,32 @@
-.notesPage {
-    display: flex;
-    padding: 0;
-    background: none;
-    flex: 1;
-}
-
 .editorContainer {
     flex: 1;
     width: calc(100vw - 35rem);
     position: relative;
+}
 
+.titleContainer {
+    background: var(--light-card-color);
+    display: flex;
+    align-items: center;
+    gap: 5px;
+    font-size: 1.8rem;
+    padding: 0.5rem;
+}
+
+.title {
+    padding: 4px 8px;
+    font-size: 1.8rem;
+    width: fit-content;
+    border: none;
+    outline: none;
+    background: var(--light-card-color);
+    color: var(--font-color-dark);
+    width: 100%;
 }
 
 .editorUtils {
     border-top: 1px solid var(--border-color);
-    padding: 0.5rem;
+    padding: 1rem 0.5rem;
     position: absolute;
     bottom: 0;
     width: 100%;
@@ -32,6 +44,7 @@
     box-shadow: var(--light-shadow);
     width: min-content;
     padding: 0.5rem;
+    border-radius: 10px;
 }
 
 .tags,
