+++ src/components/index.js
@@ -3,3 +3,4 @@ export { Toast } from "./Toast/Toast";
 export { Navigation } from "./Navigation/Navigation";
 export { Card } from "./Card/Card";
 export { Sidebar } from "./Sidebar/Sidebar";
+export { EditorContainer } from "./EditorContainer.js/EditorContainer";
